pyproject.toml
[tool.poetry]
name = "beta"
version = "0.1.2"
description = "Backend for ML dating app"
authors = ["Epanov Pavel <epanov.pavelm@gmail.com>", "Yunakov Ilia <iyunakov@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
async-sqlalchemy = "^1.0.0"
dishka = "^1.2.0"
gunicorn = "^22.0.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.10"
mypy = "^1.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
src = ["src"]
exclude = [
    "src/beta/infrastructure/data_access/migrations/*",
    "src/beta/infrastructure/data_access/models/__init__.py",
          ]

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
known_first_party = ["beta"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true